[DOKU] TA5: Wireshark-Regel dokumentiert

+++ TA5-Tracking/TA5-Trackingsystem/src/main.cpp
@@ -292,7 +292,7 @@ int main(int argc, const char** argv) {
        * Ausgabe und Abbruch
        */
       cout << "x " << (int) objectPos3D.x << "\ty " << (int) objectPos3D.y << "\tz " << (int) objectPos3D.z;
-      cout << "\t\t" << "Abstand Triangulation: " << triangulationMinDistance;
+      cout << "\t\t" << "Abstand Triangulation: " << (int)triangulationMinDistance;
       cout << "\t\t" << "Fehlercode: ";
       printf("0x%2x", positionDataErrorCode);
       cout << endl;
